Refactor landing page and app routes
- Updated routing to replace the landing page with a new app route.
- Created new app.tsx and app.module.scss files for the app layout.
- Removed old landing styles and components, consolidating into the new structure.
- Enhanced the home page with a microphone icon and improved styling.
