+++ web/app/routes.ts
@@ -2,5 +2,5 @@ import { type RouteConfig, index, route } from "@react-router/dev/routes";
 
 export default [
   index("routes/home.tsx"),
-  route("/landing", "routes/landing.tsx"),
+  route("/app", "routes/app.tsx"),
   route("/favicon.svg", "favicon.tsx")] satisfies RouteConfig;
