+++ web/app/routes/app.module.scss
@@ -1,0 +1,50 @@
+@use "../defaults.scss" as *;
+
+.home {
+  width: 100wv;
+  height: 100vh;
+}
+
+.scratch {
+  position: fixed;
+  top: 0;
+  left: 0;
+  border: solid 1px $color-fuchsia-100;
+  width: calc(100vw - 0.2rem);
+  height: calc(100vh - 0.2rem);
+  display: flex;
+  flex-direction: row;
+  flex-wrap: nowrap;
+  overflow: hidden;
+}
+
+.effort {
+  //border: solid 1px $color-fuchsia-100;
+  width: 20rem;
+  padding-top: 6rem;
+  display: flex;
+  flex-direction: column;
+  margin-left: 2.5rem;
+}
+
+.output {
+  //border: solid 1px $color-fuchsia-200;
+  width: calc(100vw - 22.5rem);
+  padding-top: 6rem;
+  padding-right: 2.5rem;
+}
+
+.talking {
+  color: red;
+}
+
+.editor {
+  max-width: 1280px;
+  width: calc(100vw - 5rem);
+  left: 1rem;
+}
+
+.voice {
+  width: 640px;
+  left: calc(50px + 2rem);
+}
